Break ingest source watcher into its own class to address complexity
